requirements.txt
accounts==0.2.4
alembic==1.7.5
amqp==2.6.1
anyio==3.4.0
arrow==1.2.1
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
bs4==0.0.1
cachetools==4.2.4
celery==4.4.7
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==7.1.2
colorama==0.4.4
config==0.5.1
cryptography==36.0.1
cssselect==1.1.0
cssutils==2.3.0
defusedxml==0.7.1
Django==3.2.8
django-allauth==0.47.0
django-appconf==1.0.5
django-compressor==3.1
django-crispy-forms==1.13.0
django-libsass==0.7
django-paypal==1.1.2
Django-Verify-Email==1.0.9
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
emails==0.5.15
fastapi==0.70.1
google-api-python-client==1.6.4
google-auth==2.3.3
greenlet==1.1.2
gunicorn==20.1.0
h11==0.9.0
httplib2==0.20.2
idna==3.3
inflect==5.3.0
install==1.3.5
Jinja2==2.11.3
joblib==1.1.0
jsonpickle==0.9.6
keras==2.7.0
kombu==4.6.11
libsass==0.21.0
lxml==4.7.1
Mako==1.1.6
MarkupSafe==2.0.1
mathparse==0.1.2
more-itertools==8.12.0
mysql-connector-python==8.0.28
nltk==3.6.6
numpy==1.21.5
oauth2client==4.1.2
oauthlib==3.1.1
opencv-python==4.5.5.62
packaging==21.3
pandas==1.3.5
parts==1.2.2
passlib==1.7.4
paypal==1.2.5
Pillow==9.0.0
Pint==0.18
pluggy==0.13.1
premailer==3.10.0
prettytable==2.5.0
products==0.3.2
protobuf==3.19.4
psycopg2-binary==2.9.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
PyJWT==2.3.0
pymongo==3.12.3
pyparsing==3.0.6
PyPDF2==1.26.0
PySocks==1.7.1
pyTelegramBotAPI==4.2.2
pytest==5.4.3
python-dateutil==2.7.5
python-jose==3.3.0
python-math==0.0.1
python-multipart==0.0.5
python3-openid==3.2.0
pytz==2021.3
PyYAML==3.13
random2==1.0.1
raven==6.10.0
rcssmin==1.1.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rjsmin==1.2.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.4.29
sqlparse==0.4.2
starlette==0.16.0
telebot==0.0.4
tenacity==6.3.1
tickets==0.3.0
tokens==0.0.3
tqdm==4.62.3
tweepy==3.8.0
typing_extensions==4.0.1
uritemplate==3.0.1
urllib3==1.26.7
uvicorn==0.11.8
vine==1.3.0
wcwidth==0.2.5
websockets==8.1
